CLN: split off frozen (immutable) data structures into pandas/indexes/frozen.py
