+++ pandas/compat/pickle_compat.py
@@ -52,12 +52,40 @@ def load_reduce(self):
 
     stack[-1] = value
 
+
+# if classes are moved, provide compat here
+_class_locations_map = {
+
+    # 15477
+    ('pandas.core.base', 'FrozenNDArray'): ('pandas.indexes.frozen', 'FrozenNDArray'),
+    ('pandas.core.base', 'FrozenList'): ('pandas.indexes.frozen', 'FrozenList')
+    }
+
+
+# our Unpickler sub-class to override methods and some dispatcher
+# functions for compat
+
 if compat.PY3:
     class Unpickler(pkl._Unpickler):
-        pass
+
+        def find_class(self, module, name):
+            # override superclass
+            key = (module, name)
+            module, name = _class_locations_map.get(key, key)
+            return super(Unpickler, self).find_class(module, name)
+
 else:
+
     class Unpickler(pkl.Unpickler):
-        pass
+
+        def find_class(self, module, name):
+            # override superclass
+            key = (module, name)
+            module, name = _class_locations_map.get(key, key)
+            __import__(module)
+            mod = sys.modules[module]
+            klass = getattr(mod, name)
+            return klass
 
 Unpickler.dispatch = copy.copy(Unpickler.dispatch)
 Unpickler.dispatch[pkl.REDUCE[0]] = load_reduce
@@ -75,8 +103,6 @@ def load_newobj(self):
 
     self.stack[-1] = obj
 Unpickler.dispatch[pkl.NEWOBJ[0]] = load_newobj
-
-# py3 compat
 
 
 def load_newobj_ex(self):
